@@ -6,14 +6,12 @@
 #
 # main author: Nils Blach
 
-from __future__ import annotations
-
 import asyncio
 import itertools
 import logging
 from abc import ABC, abstractmethod
 from enum import Enum
-from typing import Callable, Iterator, Optional, Union
+from typing import Callable, Iterator, Optional
 
 from graph_of_thoughts.language_models import AbstractLanguageModel
 from graph_of_thoughts.operations.thought import Thought
@@ -55,8 +53,8 @@ class Operation(ABC):
         """
         self.logger: logging.Logger = logging.getLogger(self.__class__.__name__)
         self.id: int = next(Operation._ids)
-        self.predecessors: "list[list[Operation]]" = []
-        self.successors: "list[list[Operation]]" = []
+        self.predecessors: list[Operation] = []
+        self.successors: list[Operation] = []
         self.executed: bool = False
 
     def can_be_executed(self) -> bool:
@@ -68,14 +66,14 @@ class Operation(ABC):
         """
         return all(predecessor.executed for predecessor in self.predecessors)
 
-    def get_previous_thoughts(self) -> "list[list[Thought]]":
+    def get_previous_thoughts(self) -> list[Thought]:
         """
         Iterates over all predecessors and aggregates their thoughts.
 
         :return: A list of all thoughts from the predecessors.
-        :rtype: "list[list[Thought]]"
+        :rtype: list[Thought]
         """
-        previous_thoughts: "list[list[Thought]]" = [
+        previous_thoughts: list[Thought] = [
             thought
             for predecessor in self.predecessors
             for thought in predecessor.get_thoughts()
@@ -83,22 +81,22 @@ class Operation(ABC):
 
         return previous_thoughts
 
-    def add_predecessor(self, operation: Operation) -> None:
+    def add_predecessor(self, operation: "Operation") -> None:
         """
         Add a preceding operation and update the relationships.
 
         :param operation: The operation to be set as a predecessor.
-        :type operation: Operation
+        :type operation: "Operation"
         """
         self.predecessors.append(operation)
         operation.successors.append(self)
 
-    def add_successor(self, operation: Operation) -> None:
+    def add_successor(self, operation: "Operation") -> None:
         """
         Add a succeeding operation and update the relationships.
 
         :param operation: The operation to be set as a successor.
-        :type operation: Operation
+        :type operation: "Operation"
         """
         self.successors.append(operation)
         operation.predecessors.append(self)
@@ -145,13 +143,13 @@ class Operation(ABC):
         pass
 
     @abstractmethod
-    def get_thoughts(self) -> "list[list[Thought]]":
+    def get_thoughts(self) -> list[Thought]:
         """
         Abstract method to retrieve the thoughts associated with the operation.
         This should be implemented in derived classes.
 
         :return: List of associated thoughts.
-        :rtype: "list[list[Thought]]"
+        :rtype: list[Thought]
         """
         pass
 
@@ -168,8 +166,8 @@ class Score(Operation):
         num_samples: int = 1,
         combined_scoring: bool = False,
         scoring_function: Callable[
-            [Union["list[list[dict]]", dict]], Union["list[list[float]]", float]
-        ] = None,
+            [list[dict] | dict], list[float] | float
+        ] | None = None,
     ) -> None:
         """
         Initializes a new Score operation.
@@ -185,17 +183,17 @@ class Score(Operation):
         super().__init__()
         self.num_samples: int = num_samples
         self.combined_scoring: bool = combined_scoring
-        self.thoughts: "list[list[Thought]]" = []
+        self.thoughts: list[Thought] = []
         self.scoring_function: Callable[
-            [Union["list[list[dict]]", dict]], Union["list[list[float]]", float]
-        ] = scoring_function
+            [list[dict] | dict], list[float] | float
+        ] | None = scoring_function
 
-    def get_thoughts(self) -> "list[list[Thought]]":
+    def get_thoughts(self) -> list[Thought]:
         """
         Returns the thoughts associated with the operation.
 
         :return: List of scored thoughts.
-        :rtype: "list[list[Thought]]"
+        :rtype: list[Thought]
         """
         return self.thoughts
 
@@ -216,7 +214,7 @@ class Score(Operation):
         :param kwargs: Additional parameters for execution.
         :raises AssertionError: If operation has no predecessors.
         """
-        previous_thoughts: "list[list[Thought]]" = self.get_previous_thoughts()
+        previous_thoughts: list[Thought] = self.get_previous_thoughts()
 
         assert (
             len(self.predecessors) > 0
@@ -283,7 +281,7 @@ class ValidateAndImprove(Operation):
         num_samples: int = 1,
         improve: bool = True,
         num_tries: int = 3,
-        validate_function: Callable[[Dict], bool] = None,
+        validate_function: Callable[[dict], bool] | None = None,
     ) -> None:
         """
         Initializes a new ValidateAndImprove operation.
@@ -301,15 +299,15 @@ class ValidateAndImprove(Operation):
         self.num_samples: int = num_samples
         self.improve: bool = improve
         self.num_tries: int = num_tries
-        self.validate_function: Callable[[Dict], bool] = validate_function
-        self.thoughts: "list[list[list[Thought]]"] = []
+        self.validate_function: Callable[[dict], bool] | None = validate_function
+        self.thoughts: list[list[Thought]] = []
 
-    def get_thoughts(self) -> "list[list[Thought]]":
+    def get_thoughts(self) -> list[Thought]:
         """
         Returns the list of final thoughts, after validation and improvement.
 
         :return: List of final validated and improved thoughts.
-        :rtype: "list[list[Thought]]"
+        :rtype: list[Thought]
         """
         return [thought_list[-1] for thought_list in self.thoughts]
 
@@ -330,7 +328,7 @@ class ValidateAndImprove(Operation):
         :param kwargs: Additional parameters for execution.
         :raises AssertionError: If operation has no predecessors.
         """
-        previous_thoughts: "list[list[Thought]]" = self.get_previous_thoughts()
+        previous_thoughts: list[Thought] = self.get_previous_thoughts()
 
         assert (
             len(self.predecessors) > 0
@@ -414,14 +412,14 @@ class Generate(Operation):
         super().__init__()
         self.num_branches_prompt: int = num_branches_prompt
         self.num_branches_response: int = num_branches_response
-        self.thoughts: "list[list[Thought]]" = []
+        self.thoughts: list[Thought] = []
 
-    def get_thoughts(self) -> "list[list[Thought]]":
+    def get_thoughts(self) -> list[Thought]:
         """
         Returns the thoughts associated with the operation.
 
         :return: List of generated thoughts.
-        :rtype: "list[list[Thought]]"
+        :rtype: list[Thought]
         """
         return self.thoughts
 
@@ -441,7 +439,7 @@ class Generate(Operation):
         :type parser: Parser
         :param kwargs: Additional parameters for execution.
         """
-        previous_thoughts: "list[list[Thought]]" = self.get_previous_thoughts()
+        previous_thoughts: list[Thought] = self.get_previous_thoughts()
 
         if len(previous_thoughts) == 0 and len(self.predecessors) > 0:
             return
@@ -494,14 +492,14 @@ class Improve(Operation):
         Initializes a new Improve operation.
         """
         super().__init__()
-        self.thoughts: "list[list[Thought]]" = []
+        self.thoughts: list[Thought] = []
 
-    def get_thoughts(self) -> "list[list[Thought]]":
+    def get_thoughts(self) -> list[Thought]:
         """
         Returns the thoughts associated with the operation after improvement.
 
         :return: List of improved thoughts.
-        :rtype: "list[list[Thought]]"
+        :rtype: list[Thought]
         """
         return self.thoughts
 
@@ -521,7 +519,7 @@ class Improve(Operation):
         :param kwargs: Additional parameters for execution.
         :raises AssertionError: If operation has no predecessors.
         """
-        previous_thoughts: "list[list[Thought]]" = self.get_previous_thoughts()
+        previous_thoughts: list[Thought] = self.get_previous_thoughts()
 
         assert len(self.predecessors) > 0, "Needs at least one predecessor"
 
@@ -553,15 +551,15 @@ class Aggregate(Operation):
         :type num_responses: int
         """
         super().__init__()
-        self.thoughts: "list[list[Thought]]" = []
+        self.thoughts: list[Thought] = []
         self.num_responses: int = num_responses
 
-    def get_thoughts(self) -> "list[list[Thought]]":
+    def get_thoughts(self) -> list[Thought]:
         """
         Returns the thoughts associated with the operation after aggregation.
 
         :return: List of aggregated thoughts.
-        :rtype: "list[list[Thought]]"
+        :rtype: list[Thought]
         """
         return self.thoughts
 
@@ -585,13 +583,13 @@ class Aggregate(Operation):
             len(self.predecessors) >= 1
         ), "Aggregate operation must have at least one predecessor"
 
-        previous_thoughts: "list[list[Thought]]" = self.get_previous_thoughts()
+        previous_thoughts: list[Thought] = self.get_previous_thoughts()
 
         if len(previous_thoughts) == 0:
             return
 
         # applied in order of score
-        base_state: Dict = {}
+        base_state: dict = {}
         for thought in sorted(previous_thoughts, key=lambda thought: thought.score):
             base_state = {**base_state, **thought.state}
 
@@ -635,18 +633,18 @@ class KeepBestN(Operation):
         self.n: int = n
         assert self.n > 0, "KeepBestN operation must keep at least one thought"
         self.higher_is_better: bool = higher_is_better
-        self.thoughts: "list[list[Thought]]" = []
+        self.thoughts: list[Thought] = []
 
-    def get_best_n(self) -> "list[list[Thought]]":
+    def get_best_n(self) -> list[Thought]:
         """
         Returns the best N thoughts from the predecessors based on their score.
 
         :return: List of best N thoughts.
-        :rtype: "list[list[Thought]]"
+        :rtype: list[Thought]
         :raises AssertionError: If not all predecessors have been executed.
         :raises AssertionError: If not all thoughts have been scored.
         """
-        previous_thoughts: "list[list[Thought]]" = self.get_previous_thoughts()
+        previous_thoughts: list[Thought] = self.get_previous_thoughts()
         assert all(
             previous_thought.scored for previous_thought in previous_thoughts
         ), "Not all thoughts have been scored"
@@ -672,12 +670,12 @@ class KeepBestN(Operation):
                 reverse=self.higher_is_better,
             )[: self.n]
 
-    def get_thoughts(self) -> "list[list[Thought]]":
+    def get_thoughts(self) -> list[Thought]:
         """
         Returns the thoughts kept by the operation.
 
         :return: List of kept thoughts.
-        :rtype: "list[list[Thought]]"
+        :rtype: list[Thought]
         """
         return self.thoughts
 
@@ -726,14 +724,14 @@ class KeepValid(Operation):
         Initializes a new KeepValid operation.
         """
         super().__init__()
-        self.thoughts: "list[list[Thought]]" = []
+        self.thoughts: list[Thought] = []
 
-    def get_thoughts(self) -> "list[list[Thought]]":
+    def get_thoughts(self) -> list[Thought]:
         """
         Returns the thoughts kept by the operation.
 
         :return: List of kept thoughts.
-        :rtype: "list[list[Thought]]"
+        :rtype: list[Thought]
         """
         return self.thoughts
 
@@ -757,7 +755,7 @@ class KeepValid(Operation):
             len(self.predecessors) >= 1
         ), "KeepValid operation must have at least one predecessor"
 
-        self.thoughts: "list[list[Thought]]" = [
+        self.thoughts: list[Thought] = [
             Thought.from_thought(thought)
             for thought in self.get_previous_thoughts()
             if not thought.validated or thought.valid
@@ -785,7 +783,7 @@ class GroundTruth(Operation):
 
     operation_type: OperationType = OperationType.ground_truth_evaluator
 
-    def __init__(self, ground_truth_evaluator: Callable[[Dict], bool]) -> None:
+    def __init__(self, ground_truth_evaluator: Callable[[dict], bool]) -> None:
         """
         Initializes a new GroundTruth operation.
 
@@ -793,8 +791,8 @@ class GroundTruth(Operation):
         :type ground_truth_evaluator: A function that takes a thought state and returns a boolean.
         """
         super().__init__()
-        self.ground_truth_evaluator: Callable[[Dict], bool] = ground_truth_evaluator
-        self.thoughts: "list[list[Thought]]" = []
+        self.ground_truth_evaluator: Callable[[dict], bool] = ground_truth_evaluator
+        self.thoughts: list[Thought] = []
 
     def get_thoughts(self) -> "list[list[Thought]]":
         """
@@ -1080,7 +1078,7 @@ class BatchGenerate(Operation):
 
     def _run_async_batch_safely(
         self, lm: AbstractLanguageModel, prompts: "list[list[str]]"
-    ) -> "list[list[Dict]]":
+    ) -> list[list[dict]]:
         """
         Safely run batch processing with optimized event loop management.
 
@@ -1089,7 +1087,7 @@ class BatchGenerate(Operation):
         :param prompts: List of prompts to process.
         :type prompts: "list[list[str]]"
         :return: List of responses.
-        :rtype: "list[list[Dict]]"
+        :rtype: list[list[dict]]
         :raises RuntimeError: If called from within an async context
         """
 
@@ -1114,7 +1112,7 @@ class BatchGenerate(Operation):
 
     async def _run_batch_async(
         self, lm: AbstractLanguageModel, prompts: "list[list[str]]"
-    ) -> "list[list[Dict]]":
+    ) -> list[list[dict]]:
         """
         Run batch processing asynchronously.
 
@@ -1123,7 +1121,7 @@ class BatchGenerate(Operation):
         :param prompts: List of prompts to process.
         :type prompts: "list[list[str]]"
         :return: List of responses.
-        :rtype: "list[list[Dict]]"
+        :rtype: list[list[dict]]
         """
         async with lm:  # Use async context manager
             return await lm.query_batch(
@@ -1189,8 +1187,8 @@ class BatchScore(Operation):
         num_samples: int = 1,
         combined_scoring: bool = False,
         scoring_function: Callable[
-            [Union["list[list[Dict]]", Dict]], Union["list[list[float]]", float]
-        ] = None,
+            [list[dict] | dict], list[float] | float
+        ] | None = None,
         max_concurrent: Optional[int] = None,
         batch_size: Optional[int] = None,
         enable_batch_processing: bool = True,
@@ -1216,8 +1214,8 @@ class BatchScore(Operation):
         self.num_samples: int = num_samples
         self.combined_scoring: bool = combined_scoring
         self.scoring_function: Callable[
-            [Union["list[list[Dict]]", Dict]], Union["list[list[float]]", float]
-        ] = scoring_function
+            [list[dict] | dict], list[float] | float
+        ] | None = scoring_function
         self.max_concurrent: Optional[int] = max_concurrent
         self.batch_size: Optional[int] = batch_size
         self.enable_batch_processing: bool = enable_batch_processing
@@ -1386,7 +1384,7 @@ class BatchScore(Operation):
 
     def _run_async_batch_safely(
         self, lm: AbstractLanguageModel, prompts: "list[list[str]]"
-    ) -> "list[list[Dict]]":
+    ) -> list[list[dict]]:
         """
         Safely run batch processing with optimized event loop management.
 
@@ -1395,7 +1393,7 @@ class BatchScore(Operation):
         :param prompts: List of prompts to process.
         :type prompts: "list[list[str]]"
         :return: List of responses.
-        :rtype: "list[list[Dict]]"
+        :rtype: list[list[dict]]
         :raises RuntimeError: If called from within an async context
         """
 
@@ -1420,7 +1418,7 @@ class BatchScore(Operation):
 
     async def _run_batch_async(
         self, lm: AbstractLanguageModel, prompts: "list[list[str]]"
-    ) -> "list[list[Dict]]":
+    ) -> list[list[dict]]:
         """
         Run batch processing asynchronously for scoring.
 
@@ -1429,7 +1427,7 @@ class BatchScore(Operation):
         :param prompts: List of prompts to process.
         :type prompts: "list[list[str]]"
         :return: List of responses.
-        :rtype: "list[list[Dict]]"
+        :rtype: list[list[dict]]
         """
         async with lm:  # Use async context manager
             return await lm.query_batch(
@@ -1554,7 +1552,7 @@ class BatchAggregate(Operation):
             return
 
         # applied in order of score
-        base_state: Dict = {}
+        base_state: dict = {}
         for thought in sorted(previous_thoughts, key=lambda thought: thought.score):
             base_state = {**base_state, **thought.state}
 
@@ -1585,8 +1583,8 @@ class BatchAggregate(Operation):
         lm: AbstractLanguageModel,
         prompter: Prompter,
         parser: Parser,
-        previous_thought_states: "list[list[Dict]]",
-        base_state: Dict,
+        previous_thought_states: list[dict],
+        base_state: dict,
     ) -> None:
         """
         Execute aggregation using batch processing for improved performance.
@@ -1598,9 +1596,9 @@ class BatchAggregate(Operation):
         :param parser: The parser for parsing responses.
         :type parser: Parser
         :param previous_thought_states: List of thought states to aggregate.
-        :type previous_thought_states: "list[list[Dict]]"
+        :type previous_thought_states: list[dict]
         :param base_state: Base state for new thoughts.
-        :type base_state: Dict
+        :type base_state: dict
         """
         self.logger.info(f"Using batch aggregation for {self.num_responses} responses")
 
@@ -1649,7 +1647,7 @@ class BatchAggregate(Operation):
 
     def _run_async_batch_safely(
         self, lm: AbstractLanguageModel, prompts: "list[list[str]]"
-    ) -> "list[list[Dict]]":
+    ) -> list[list[dict]]:
         """
         Safely run batch processing with optimized event loop management.
 
@@ -1658,7 +1656,7 @@ class BatchAggregate(Operation):
         :param prompts: List of prompts to process.
         :type prompts: "list[list[str]]"
         :return: List of responses.
-        :rtype: "list[list[Dict]]"
+        :rtype: list[list[dict]]
         :raises RuntimeError: If called from within an async context
         """
 
@@ -1683,7 +1681,7 @@ class BatchAggregate(Operation):
 
     async def _run_batch_async(
         self, lm: AbstractLanguageModel, prompts: "list[list[str]]"
-    ) -> "list[list[Dict]]":
+    ) -> list[list[dict]]:
         """
         Run batch processing asynchronously for aggregation.
 
@@ -1692,7 +1690,7 @@ class BatchAggregate(Operation):
         :param prompts: List of prompts to process.
         :type prompts: "list[list[str]]"
         :return: List of responses.
-        :rtype: "list[list[Dict]]"
+        :rtype: list[list[dict]]
         """
         async with lm:  # Use async context manager
             return await lm.query_batch(
@@ -1704,8 +1702,8 @@ class BatchAggregate(Operation):
         lm: AbstractLanguageModel,
         prompter: Prompter,
         parser: Parser,
-        previous_thought_states: "list[list[Dict]]",
-        base_state: Dict,
+        previous_thought_states: list[dict],
+        base_state: dict,
     ) -> None:
         """
         Execute aggregation using sequential processing (fallback method).
@@ -1717,9 +1715,9 @@ class BatchAggregate(Operation):
         :param parser: The parser for parsing responses.
         :type parser: Parser
         :param previous_thought_states: List of thought states to aggregate.
-        :type previous_thought_states: "list[list[Dict]]"
+        :type previous_thought_states: list[dict]
         :param base_state: Base state for new thoughts.
-        :type base_state: Dict
+        :type base_state: dict
         """
         self.logger.info(
             f"Using sequential aggregation for {self.num_responses} responses"
