@@ -11,7 +11,8 @@ import datetime
 import json
 import logging
 import os
-from typing import Callable, Dict, List, Union
+# Type annotations use built-in types for Python 3.12+ compatibility
+from typing import Callable
 
 from graph_of_thoughts import controller, language_models, operations, parser, prompter
 
@@ -130,7 +131,7 @@ Only output the final 4 lists in the following format without any additional tex
 
 <Example>
 Input: [3, 1, 9, 3, 7, 5, 5, 4, 8, 1, 5, 3, 3, 2, 3, 0, 9, 7, 2, 2, 4, 4, 8, 5, 0, 8, 7, 3, 3, 8, 7, 0, 9, 5, 1, 6, 7, 6, 8, 9, 0, 3, 0, 6, 3, 4, 8, 0, 6, 9, 8, 4, 1, 2, 9, 0, 4, 8, 8, 9, 9, 8, 5, 9]
-Output: 
+Output:
 {{
     "List 1": [3, 1, 9, 3, 7, 5, 5, 4, 8, 1, 5, 3, 3, 2, 3, 0],
     "List 2": [9, 7, 2, 2, 4, 4, 8, 5, 0, 8, 7, 3, 3, 8, 7, 0],
